ipaccess: Properly resolve 'line' in error case
If a write on an IPA file descriptor fails, then we call
ipaccess_drop().  However, only in the BSC side the assumption that
bfd->data == e1inp_line holds true. On the BTS side, ofd->data
references to the ipa_client_conn structure.  In order to avoid the
problem, call ipaccess_drop() with an explicit reference to the line in
which the link was dropped.